[IMP] alc_all: Drop column only if exists

+++ odoo/addons/alc_all/migrations/16.0.2.10.2/end-migrate.py
@@ -2,11 +2,14 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 from openupgradelib import openupgrade
 
+from odoo.tools import sql
+
 
 def migrate(cr, version):
     """Remove obsolete FK."""
 
-    query = """
-        alter table stock_move_line drop column additional_move_id ;
-      """
-    openupgrade.logged_query(cr, query)
+    if sql.column_exists(cr, "stock_move_line", "additional_move_id"):
+        query = """
+          alter table stock_move_line drop column additional_move_id ;
+        """
+        openupgrade.logged_query(cr, query)
